Fix nested template and data model issues
The data structure is interesting here - we have one JSON structure that
gets returned by the SuitesRoute model hook. The SuiteRoute model hook
also calls store.find, but the `suite` model data is already present in
the store, so it doesn't query the Fixture data. This is the same as
using modelFor('suites').filter(), which is what I do in the
SuiteTestRoute.

I also changed the `test` resource to a nested route, so we have
`suite.test` (route name in the router and in link-to helpers) and a
SuiteTestRoute instead of a TestRoute, as well as a `suite/test`
template id.

Finally, there was some tangle with the table markup and nested each
helpers which was causing the bulk of the visible breakage after
clicking back and forth on a few tests.